+++ vite.config.ts
@@ -9,6 +9,7 @@ export default defineConfig({
     alias: {
       '@/assets': path.resolve(__dirname, './src/assets'),
       '@/components': path.resolve(__dirname, './src/components'),
+      '@/consts': path.resolve(__dirname, './src/consts'),
       '@/containers': path.resolve(__dirname, './src/containers'),
       '@/redux': path.resolve(__dirname, './src/redux'),
       '@/services': path.resolve(__dirname, './src/services'),